+++ commanderbot_ext/_lib/cog_guild_state.py
@@ -1,0 +1,39 @@
+from dataclasses import dataclass, field
+
+from commanderbot_lib.logging import Logger, get_logger
+from discord import Guild
+from discord.ext.commands import Bot, Cog
+
+
+@dataclass
+class CogGuildState:
+    """
+    Encapsulates the state and logic of a particular cog, at the guild level.
+
+    This is a thin abstraction that affords us two main conveniences:
+    1. keeping guild-based logic separate from global logic; and
+    2. keeping the state of each guild isolated from one another.
+
+    Attributes
+    -----------
+    bot
+        The bot/client instance the cog is attached to.
+    cog
+        The cog instance this state is attached to.
+    guild
+        The guild instance being managed.
+    log
+        A logger named in a uniquely identifiable way.
+    """
+
+    bot: Bot
+    cog: Cog
+    guild: Guild
+
+    log: Logger = field(init=False)
+
+    def __post_init__(self):
+        self.log = get_logger(
+            f"{self.cog.qualified_name}@{self.guild}"
+            + f" ({self.__class__.__name__}#{id(self)})"
+        )