+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Added `pack` command that runs a `beet` build and uses `lectern`
+
 ## [0.6.0] - 2021-01-08
 
 ### Added