+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.8.0] - 2021-05-08
 
 ### Added
 
@@ -77,7 +77,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Implemented `status` and `faq` extensions as an exercise for developing `commanderbot-lib`
 
-[unreleased]: https://github.com/CommanderBot-Dev/commanderbot-ext/compare/v0.7.0...HEAD
+[unreleased]: https://github.com/CommanderBot-Dev/commanderbot-ext/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/CommanderBot-Dev/commanderbot-ext/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/CommanderBot-Dev/commanderbot-ext/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/CommanderBot-Dev/commanderbot-ext/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/CommanderBot-Dev/commanderbot-ext/compare/v0.4.0...v0.5.0